+++ activerecord/lib/rails/generators/active_record/model/model_generator.rb
@@ -5,7 +5,7 @@ require "rails/generators/active_record"
 module ActiveRecord
   module Generators # :nodoc:
     class ModelGenerator < Base # :nodoc:
-      argument :attributes, type: :array, default: [], banner: "field[:type][:index] field[:type][:index]"
+      argument :attributes, type: :array, default: [], banner: "field[:type][:required][:index] field[:type][:required][:index]"
 
       check_class_collision