+++ railties/lib/rails/generators/generated_attribute.rb
@@ -5,32 +5,17 @@ require "active_support/time"
 module Rails
   module Generators
     class GeneratedAttribute # :nodoc:
-      INDEX_OPTIONS = %w(index uniq)
-      UNIQ_INDEX_OPTIONS = %w(uniq)
-
       attr_accessor :name, :type
       attr_reader   :attr_options
       attr_writer   :index_name
 
       class << self
         def parse(column_definition)
-          name, type, has_index = column_definition.split(":")
-
-          # if user provided "name:index" instead of "name:string:index"
-          # type should be set blank so GeneratedAttribute's constructor
-          # could set it to :string
-          has_index, type = type, nil if INDEX_OPTIONS.include?(type)
-
-          type, attr_options = *parse_type_and_options(type)
+          name, *definition = column_definition.split(":")
+          type, attr_options = *parse_type_and_options(definition)
           type = type.to_sym if type
 
-          if type && reference?(type)
-            if UNIQ_INDEX_OPTIONS.include?(has_index)
-              attr_options[:index] = { unique: true }
-            end
-          end
-
-          new(name, type, has_index, attr_options)
+          new(name, type, attr_options)
         end
 
         def reference?(type)
@@ -41,29 +26,49 @@ module Rails
 
           # parse possible attribute options like :limit for string/text/binary/integer, :precision/:scale for decimals or :polymorphic for references/belongs_to
           # when declaring options curly brackets should be used
-          def parse_type_and_options(type)
+          def parse_type_and_options(definition)
+            type, *provided_options = definition
+            options = {}
+
+            # if user provided "name:index" instead of "name:string:index" or
+            # provided "name:required" instead of "name:string:required",
+            # type should be set blank so GeneratedAttribute's constructor
+            # could set it to :string
+            if %w(required index uniq).include?(type)
+              provided_options << type
+              type = nil
+            end
+
+            if provided_options.include?("required")
+              options[:required] = true
+            end
+
+            if provided_options.include?("uniq")
+              options[:index] = { unique: true }
+            elsif provided_options.include?("index")
+              options[:index] = true
+            end
+
             case type
             when /(string|text|binary|integer)\{(\d+)\}/
-              return $1, limit: $2.to_i
+              return $1, options.merge(limit: $2.to_i)
             when /decimal\{(\d+)[,.-](\d+)\}/
-              return :decimal, precision: $1.to_i, scale: $2.to_i
+              return :decimal, options.merge(precision: $1.to_i, scale: $2.to_i)
             when /(references|belongs_to)\{(.+)\}/
               type = $1
-              provided_options = $2.split(/[,.-]/)
-              options = Hash[provided_options.map { |opt| [opt.to_sym, true] }]
-              return type, options
+              provided_attr_options = $2.split(/[,.-]/)
+              attr_options = Hash[provided_attr_options.map { |opt| [opt.to_sym, true] }]
+              return type, options.merge(attr_options)
             else
-              return type, {}
+              return type, options
             end
           end
       end
 
-      def initialize(name, type = nil, index_type = false, attr_options = {})
-        @name           = name
-        @type           = type || :string
-        @has_index      = INDEX_OPTIONS.include?(index_type)
-        @has_uniq_index = UNIQ_INDEX_OPTIONS.include?(index_type)
-        @attr_options   = attr_options
+      def initialize(name, type = nil, attr_options = {})
+        @name         = name
+        @type         = type || :string
+        @attr_options = attr_options
       end
 
       def field_type
@@ -137,11 +142,11 @@ module Rails
       end
 
       def has_index?
-        @has_index
+        !!attr_options[:index]
       end
 
       def has_uniq_index?
-        @has_uniq_index
+        attr_options[:index].is_a?(Hash) && !!attr_options.dig(:index, :unique)
       end
 
       def password_digest?
@@ -162,6 +167,10 @@ module Rails
 
       def options_for_migration
         @attr_options.dup.tap do |options|
+          if has_index?
+            options.delete(:index)
+          end
+
           if required?
             options.delete(:required)
             options[:null] = false