+++ railties/lib/rails/generators/rails/migration/migration_generator.rb
@@ -3,7 +3,7 @@
 module Rails
   module Generators
     class MigrationGenerator < NamedBase # :nodoc:
-      argument :attributes, type: :array, default: [], banner: "field[:type][:index] field[:type][:index]"
+      argument :attributes, type: :array, default: [], banner: "field[:type][:required][:index] field[:type][:required][:index]"
       hook_for :orm, required: true, desc: "ORM to be invoked"
     end
   end