+++ railties/lib/rails/generators/rails/model/model_generator.rb
@@ -7,7 +7,7 @@ module Rails
     class ModelGenerator < NamedBase # :nodoc:
       include Rails::Generators::ModelHelpers
 
-      argument :attributes, type: :array, default: [], banner: "field[:type][:index] field[:type][:index]"
+      argument :attributes, type: :array, default: [], banner: "field[:type][:required][:index] field[:type][:required][:index]"
       hook_for :orm, required: true, desc: "ORM to be invoked"
     end
   end